Move category labels to ui-category component
