+++ ui/src/utils/categories.js
@@ -9,8 +9,6 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0
  */
 
-import { msg } from 'hybrids';
-
 const colors = {
   advertising: '#cb55cd',
   audio_video_player: '#ef671e',
@@ -47,64 +45,12 @@ export const order = [
   'unknown',
 ];
 
-const labels = {
-  get advertising() {
-    return msg`Advertising | Ghostery organizes tags by category. This is one of several tag categories.`;
-  },
-  get audio_video_player() {
-    return msg`Audio/Video Player | Ghostery organizes tags by category. This is one of several tag categories.`;
-  },
-  get cdn() {
-    return msg`CDN | Ghostery organizes tags by category. This is one of several tag categories.`;
-  },
-  get comments() {
-    return msg`Comments | Ghostery organizes tags by category. This is one of several tag categories.`;
-  },
-  get consent() {
-    return msg`Consent Management | Includes trackers used for cookie consent management, allowing websites different levels of tracking user activity.`;
-  },
-  get customer_interaction() {
-    return msg`Customer Interaction | Ghostery organizes tags by category. This is one of several tag categories.`;
-  },
-  get email() {
-    return msg`Email | Ghostery organizes tags by category. This is one of several tag categories.`;
-  },
-  get essential() {
-    return msg`Essential | Ghostery organizes tags by category. This is one of several tag categories.`;
-  },
-  get extensions() {
-    return msg`Extensions | Ghostery organizes tags by category. This is one of several tag categories.`;
-  },
-  get hosting() {
-    return msg`Hosting | Ghostery organizes tags by category. This is one of several tag categories.`;
-  },
-  get misc() {
-    return msg`Miscellaneous | Ghostery organizes tags by category. This is one of several tag categories.`;
-  },
-  get pornvertising() {
-    return msg`Adult Advertising | Ghostery organizes tags by category. This is one of several tag categories.`;
-  },
-  get site_analytics() {
-    return msg`Site Analytics | Ghostery organizes tags by category. This is one of several tag categories.`;
-  },
-  get social_media() {
-    return msg`Social Media | Ghostery organizes tags by category. This is one of several tag categories.`;
-  },
-  get unknown() {
-    return msg`Unknown | Ghostery organizes tags by category. This is one of several tag categories.`;
-  },
-};
-
 export const getCategoryKey = (category) => {
   return colors[category] ? category : 'unknown';
 };
 
 export const getCategoryColor = (category) => {
   return colors[getCategoryKey(category)];
-};
-
-export const getCategoryLabel = (category) => {
-  return labels[getCategoryKey(category)];
 };
 
 export const sortCategories = (resolveCategoryName = (a) => a) => {
